# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Please enter a valid token to test the requests below..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Endpoint to login.",
        "parameters": [
          {
            "name": "userCredentials",
            "in": "body",
            "schema": {
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "maxLength": 250,
                  "example": "thang2@gmail.com"
                },
                "password": {
                  "type": "string",
                  "minLength": 7,
                  "maxLength": 250,
                  "example": "123456"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful."
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found "
          },
          "422": {
            "description": "Validation failed."
          },
          "500": {
            "description": "Error Internal Server"
          }
        }
      }
    },
    "/auth/create": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Endpoint to sign up an acount with email. A email with OTP will sent to your email.",
        "parameters": [
          {
            "name": "newUser",
            "in": "body",
            "schema": {
              "required": [
                "name",
                "email",
                "password"
              ],
              "properties": {
                "name": {
                  "type": "string",
                  "maxLength": 250,
                  "example": "Thang Mai."
                },
                "email": {
                  "type": "string",
                  "maxLength": 250,
                  "example": "thang@gmail.com"
                },
                "password": {
                  "type": "string",
                  "minLength": 6,
                  "maxLength": 250,
                  "example": "1234567"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not found "
          },
          "422": {
            "description": "Validation failed."
          },
          "500": {
            "description": "Error Internal Server"
          }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Endpoint to verify otp.",
        "parameters": [
          {
            "name": "verifyotp",
            "in": "body",
            "schema": {
              "required": [
                "otp",
                "email"
              ],
              "properties": {
                "opt": {
                  "type": "string",
                  "Length": 6,
                  "example": "123456"
                },
                "email": {
                  "type": "string",
                  "maxLength": 250,
                  "example": "thang@gmail.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not found "
          },
          "500": {
            "description": "Error Internal Server"
          }
        }
      }
    },
    "/auth/reset-password": {
      "put": {
        "tags": [
          "Auth"
        ],
        "description": "Endpoint reset password.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "resetLink": {
                  "example": "any"
                },
                "newPass": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not found "
          },
          "422": {
            "description": "Validation failed."
          },
          "500": {
            "description": "Error Internal Server"
          }
        }
      }
    },
    "/auth/forgot-password": {
      "put": {
        "tags": [
          "Auth"
        ],
        "description": "Endpoint forgot password and then receive a random password in email .",
        "parameters": [
          {
            "name": "reset-password",
            "in": "body",
            "schema": {
              "required": [
                "email"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "maxLength": 250,
                  "example": "thang@gmail.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found "
          },
          "500": {
            "description": "Error Internal Server"
          }
        }
      }
    },
...
  },
...
}